New folder/requirements.txt
agno>=2.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
duckduckgo-search>=6.0.0
newspaper4k<=0.9.4.1 
httpx>=0.27.0